bugfix: CADENCE must be a Double to avoid runtime error from cast.
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@929 0c594bb0-8543-0410-a25e-d8ec440261db

+++ BinaryDataSource/src/org/virbo/binarydatasource/BinaryDataSource.java
@@ -127,7 +127,7 @@ public class BinaryDataSource extends AbstractDataSource {
                     return offset + i * recSizeBytes;
                 }
             };
-            dep0ds.putProperty( QDataSet.CADENCE, recSizeBytes );
+            dep0ds.putProperty( QDataSet.CADENCE, (double)recSizeBytes );
             ds.putProperty(QDataSet.DEPEND_0, dep0ds);
         }
